refact: decouple tree building and tree enrichment
