+++ investment_tree/__init__.py
@@ -1,0 +1,11 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Created on Sat Jun  7 13:58:15 2025
+
+@author: andrefelix
+"""
+
+
+from .builder import build_tree
+from .enrichment import enrich_tree
