+++ reporting/__init__.py
@@ -1,0 +1,12 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Created on Sat Jun  7 14:46:57 2025
+
+@author: andrefelix
+"""
+
+
+from .governance_struct import assign_governance_struct_keys
+
+__all__ = ["assign_governance_struct_keys"]
